server/requirements.txt
Flask==2.0.2
Flask_Cors==3.0.10
Flask_SQLAlchemy==2.5.1
python-dotenv==1.0.0
